Bug report #133
 - PostUI: Prepravljeno da se umesto broja lajkova vraća prosečna ocena
 - PostDTO, PostOnMap: Promenjen tip podatka numberOfLikes i NumberOfLikes u double
 - LocationDAL: uočeni bagovi (treba još da se testira i sredi)
 - PostDAL: Dodata logika za sortiranje po prosečnoj oceni i uočen bug koji je sređen

+++ ProductionApp/back/DTOComponents/PostDTO.cs
@@ -8,7 +8,7 @@
         public string FullCoverImagePath { get; set; } = string.Empty;
 
         //public DateTime DateCreated { get; set; }
-        public int NumberOfLikes { get; set; } = 0;
+        public double NumberOfLikes { get; set; } = 0;
         public int NumberOfViews { get; set; } = 0;
         public bool IsLiked { get; set; } = false;
         public int Grade { get; set; } = 0;
